move edges to a different class
